requirements.txt
ultralytics==8.0.114
opencv-python==4.7.0.72
arabic-reshaper==3.0.0
python-bidi==0.4.2
deep-translator==1.11.1
pyobjc==9.2
gTTS==2.3.2